metalinter fixes
Test Plan: Ran integration test

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5005

+++ cache/elastimemcache/elastimemcache.go
@@ -185,6 +185,6 @@ func (c *Client) Set(key string, value string) error {
 	return c.memcacheClient.Set(&memcache.Item{
 		Key:        c.createCacheKey(key),
 		Value:      []byte(value),
-		Expiration: int32(c.config.TTL),
+		Expiration: c.config.TTL,
 	})
 }
